preserve and retrieve musicAlbum and genre data
