+++ data/music_data/genre.json
@@ -1,1 +1,1 @@
-[{"name":"Grime","id":74},{"name":"Grimm","id":28},{"name":"Hip-pop","id":961}]
+[{"name":"Grime","id":74},{"name":"Grimm","id":28},{"name":"Hip-pop","id":961},{"name":"Reggae","id":538}]
