+++ data/music_data/music.json
@@ -1,1 +1,1 @@
-[{"name":"Conspiracy","artist":"J_hus","id":782,"genre":"Grime","on_spotify":true,"publish_date":"15-03-2020"},{"name":"We're all together","artist":"Dave","id":212,"genre":"Grimm","on_spotify":false,"publish_date":"4/4/2021"},{"name":"Psycho","artist":"Dave","id":836,"genre":"Grime","on_spotify":true,"publish_date":"10/6/2018"},{"name":"Timeless","artist":"Davido","id":842,"genre":"Hip-pop","on_spotify":true,"publish_date":"15/04/2023"}]
+[{"name":"Conspiracy","artist":"J_hus","id":782,"genre":"Grime","on_spotify":true,"publish_date":"15-03-2020"},{"name":"We're all together","artist":"Dave","id":212,"genre":"Grimm","on_spotify":false,"publish_date":"4/4/2021"},{"name":"Psycho","artist":"Dave","id":836,"genre":"Grime","on_spotify":true,"publish_date":"10/6/2018"},{"name":"Timeless","artist":"Davido","id":842,"genre":"Hip-pop","on_spotify":true,"publish_date":"15/04/2023"},{"name":"Bob marley and the wailers","artist":"Bob marley","id":649,"genre":"Reggae","on_spotify":false,"publish_date":"13/6/1986"}]
